+++ js/admin.js
@@ -1,3 +1,9 @@
+let themes = {
+  'github': 'Github',
+  'github-dark': 'Github Dark',
+  'base16/monokai': 'Monokai',
+};
+
 app.initializers.add('server-side-highlight', function(app) {
   app.extensionData
     .for('club-1-server-side-highlight')
@@ -12,6 +18,13 @@ app.initializers.add('server-side-highlight', function(app) {
       type: 'color',
     })
     .registerSetting({
+      setting: 'club-1-server-side-highlight.light_theme_highlight_theme',
+      label: app.translator.trans('club-1-server-side-highlight.admin.light_theme_highlight_theme'),
+      type: 'select',
+      options: themes,
+      default: 'github',
+    })
+    .registerSetting({
       setting: 'club-1-server-side-highlight.dark_theme_bg_color',
       label: app.translator.trans('club-1-server-side-highlight.admin.dark_theme_bg_color'),
       type: 'color',
@@ -20,6 +33,13 @@ app.initializers.add('server-side-highlight', function(app) {
       setting: 'club-1-server-side-highlight.dark_theme_text_color',
       label: app.translator.trans('club-1-server-side-highlight.admin.dark_theme_text_color'),
       type: 'color',
+    })
+    .registerSetting({
+      setting: 'club-1-server-side-highlight.dark_theme_highlight_theme',
+      label: app.translator.trans('club-1-server-side-highlight.admin.dark_theme_highlight_theme'),
+      type: 'select',
+      options: themes,
+      default: 'github-dark',
     });
 });